requirements.txt
path.py
fire
